snip TSB links
